requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langsmith
python-dotenv
openai
pydantic
pytesseract
pdf2image 
langgraph
langchain-mcp-adapters
PyPDF2
pillow
chromadb 
langchain-core 
redis
fastapi-utilities 